ISSUE-179: Show target impacts inside the stats
