+++ app/src/main/AndroidManifest.xml
@@ -37,14 +37,6 @@
                 android:value="ar.com.tzulberti.archerytraining.activities.tournament.ViewExistingTournamentsActivity" />
         </activity>
 
-        <activity android:name="ar.com.tzulberti.archerytraining.activities.tournament.ViewAllTournamentTargetArrowActivity"
-            android:parentActivityName="ar.com.tzulberti.archerytraining.activities.MainActivity" >
-            <!-- The meta-data tag is required if you support API level 15 and lower -->
-            <meta-data
-                android:name="android.support.PARENT_ACTIVITY"
-                android:value="ar.com.tzulberti.archerytraining.activities.tournament.ViewExistingTournamentsActivity" />
-        </activity>
-
         <activity android:name="ar.com.tzulberti.archerytraining.activities.tournament.ViewSerieInformationActivity"
             android:parentActivityName="ar.com.tzulberti.archerytraining.activities.MainActivity" >
             <!-- The meta-data tag is required if you support API level 15 and lower -->
