+++ app/src/main/java/ar/com/tzulberti/archerytraining/activities/tournament/ViewTournamentSeriesActivity.java
@@ -139,19 +139,9 @@ public class ViewTournamentSeriesActivity extends BaseArcheryTrainingActivity im
 
         boolean buttonsEnabled = ! this.tournament.series.isEmpty();
         // add the buttons to delete/view charts for the tournament
-        TableRow tr1 = new TableRow(this);
-        Button viewChartsButton = new Button(this);
-        viewChartsButton.setId(Integer.MAX_VALUE - 1);
-        viewChartsButton.setText(R.string.tournament_view_all_impacts);
-        viewChartsButton.setLayoutParams(trParams);
-        viewChartsButton.setOnClickListener(this);
-        viewChartsButton.setEnabled(buttonsEnabled);
-        tr1.addView(viewChartsButton);
-        this.dataContainer.addView(tr1);
-
         TableRow tr2 = new TableRow(this);
         Button viewScoreSheetButton = new Button(this);
-        viewScoreSheetButton.setId(Integer.MAX_VALUE - 2);
+        viewScoreSheetButton.setId(Integer.MAX_VALUE - 1);
         viewScoreSheetButton.setText(R.string.tournament_view_score_sheet);
         viewScoreSheetButton.setLayoutParams(trParams);
         viewScoreSheetButton.setOnClickListener(this);
@@ -162,7 +152,7 @@ public class ViewTournamentSeriesActivity extends BaseArcheryTrainingActivity im
 
         TableRow tr3 = new TableRow(this);
         Button viewTournamentArrowStats = new Button(this);
-        viewTournamentArrowStats.setId(Integer.MAX_VALUE - 3);
+        viewTournamentArrowStats.setId(Integer.MAX_VALUE - 2);
         viewTournamentArrowStats.setText(R.string.tournament_view_arrow_stats);
         viewTournamentArrowStats.setLayoutParams(trParams);
         viewTournamentArrowStats.setOnClickListener(this);
@@ -187,16 +177,14 @@ public class ViewTournamentSeriesActivity extends BaseArcheryTrainingActivity im
         int id = v.getId();
 
 
-        if (id == Integer.MAX_VALUE - 1 || id == Integer.MAX_VALUE - 2 || id == Integer.MAX_VALUE -3 ) {
+        if (id == Integer.MAX_VALUE - 1 || id == Integer.MAX_VALUE - 2) {
             // selected the option to view all impacts for the current tournament
             Intent intent = null;
-            if (id == Integer.MAX_VALUE - 1) {
-                intent = new Intent(this, ViewAllTournamentTargetArrowActivity.class);
 
-            } else if (id == Integer.MAX_VALUE - 2) {
+            if (id == Integer.MAX_VALUE - 1) {
                 intent = new Intent(this, ViewTournamentScoreSheetActivity.class);
 
-            } else if (id == Integer.MAX_VALUE - 3) {
+            } else if (id == Integer.MAX_VALUE - 2) {
                 intent = new Intent(this, ContainerStatsActivity.class);
             }
