+++ app/src/main/res/layout/common_view_container_arrow_stats.xml
@@ -10,6 +10,28 @@
         android:layout_height="match_parent"
         android:orientation="vertical">
 
+
+        <TextView
+            android:id="@+id/tournament_series_showing"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content" />
+
+
+        <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:orientation="horizontal"
+            android:padding="20sp">
+
+            <com.edmodo.rangebar.RangeBar
+                xmlns:custom="http://schemas.android.com/apk/res-auto"
+                android:id="@+id/tournament_series_rangebar"
+                android:layout_width="match_parent"
+                android:layout_height="match_parent"
+                custom:tickCount="5"
+                />
+        </LinearLayout>
+
         <TableLayout xmlns:android="http://schemas.android.com/apk/res/android"
             android:id="@+id/tournament_stats_table_data"
             android:layout_width="match_parent"
@@ -40,6 +62,15 @@
 
         </TableLayout>
 
+        <ImageView
+            android:id="@+id/photo_view"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:src="@drawable/complete_archery_target"
+            android:scaleType="fitXY"
+            android:adjustViewBounds="true"
+            />
+
 
         <com.github.mikephil.charting.charts.LineChart
             android:id="@+id/tournament_series_chart"
@@ -55,88 +86,3 @@
             />
     </LinearLayout>
 </ScrollView>
-
-<!--
-    <?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-android:layout_width="match_parent"
-android:layout_height="match_parent"
-android:overScrollMode="always"
-android:scrollbarStyle="insideOverlay"
-android:scrollbars="vertical">
-
-
-<TableLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:id="@+id/todayTotalsTextData"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:stretchColumns="*"
-    >
-    <TableRow>
-        <TextView
-            android:padding="3dip"
-            android:text="@string/addSerieDistance" />
-
-        <TextView
-            android:padding="3dip"
-            android:text="@string/addSerieArrowsAmountTableHeader" />
-
-        <TextView
-            android:padding="3dip"
-            android:text="@string/addSerieDatetime" />
-
-        <TextView
-            android:padding="3dip"
-            android:text="@string/todayDataTotalSeries" />
-    </TableRow>
-    <TableRow>
-        <TextView
-            android:padding="3dip"
-            android:text="@string/addSerieDistance" />
-
-        <TextView
-            android:padding="3dip"
-            android:text="@string/addSerieArrowsAmountTableHeader" />
-
-        <TextView
-            android:padding="3dip"
-            android:text="@string/addSerieDatetime" />
-
-        <TextView
-            android:padding="3dip"
-            android:text="@string/todayDataTotalSeries" />
-    </TableRow>
-    <TableRow>
-        <TextView
-            android:padding="3dip"
-            android:text="@string/addSerieDistance" />
-
-        <TextView
-            android:padding="3dip"
-            android:text="@string/addSerieArrowsAmountTableHeader" />
-
-        <TextView
-            android:padding="3dip"
-            android:text="@string/addSerieDatetime" />
-
-        <TextView
-            android:padding="3dip"
-            android:text="@string/todayDataTotalSeries" />
-    </TableRow>
-</TableLayout>
-
-
-<com.github.mikephil.charting.charts.LineChart
-    android:id="@+id/tournament_series_chart"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:layout_below="@id/todayTotalsTextData" />
-
-<com.github.mikephil.charting.charts.HorizontalBarChart
-    android:id="@+id/tournament_arrows_horizontal_chart"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:layout_below="@id/tournament_series_chart"
-    />
-</RelativeLayout>
--->
