--- app/src/main/res/layout/tournament_view_tournament_series_on_target.xml
@@ -1,55 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:orientation="vertical"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-
-    <TextView
-        android:id="@+id/tournament_series_showing"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content" />
-
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal"
-        android:padding="20sp">
-
-
-
-        <com.edmodo.rangebar.RangeBar
-            xmlns:custom="http://schemas.android.com/apk/res-auto"
-            android:id="@+id/tournament_series_rangebar"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            custom:tickCount="5"
-            />
-    </LinearLayout>
-
-
-
-    <!--
-    <com.github.chrisbanes.photoview.PhotoView
-        android:id="@+id/photo_view"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:src="@drawable/complete_archery_target"
-        />
-    -->
-
-    <ImageView
-        android:id="@+id/photo_view"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:src="@drawable/complete_archery_target"
-        android:scaleType="fitXY"
-        android:adjustViewBounds="true"
-        />
-
-
-
-
-
-</LinearLayout>
